requirements.txt
flask
flask-socketio
eventlet
pyserial
